requirements.txt
telethon
pyyaml
logmagix
